requirements.txt
absl-py==0.8.0
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gast==0.3.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
Pillow==6.1.0
protobuf==3.9.1
psutil==5.6.3
pyparsing==2.4.2
pytesseract==0.3.0
python-dateutil==2.8.0
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tqdm==4.35.0
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
